+++ client/web/src/site/CodeHostScopeAlerts/CodeHostScopeAlerts.tsx
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom'
 import { DismissibleAlert } from '../../components/DismissibleAlert'
 import { githubRepoScopeRequired, gitlabAPIScopeRequired } from '../../user/settings/cloud-ga'
 
-import { useCodeHostScopeContext } from './GithubScopeProvider'
+import { useCodeHostScopeContext } from './CodeHostScopeProvider'
 
 interface Props {
     authenticatedUser: { id: string; tags: string[] } | null
@@ -17,7 +17,7 @@ export const GITLAB_SCOPE_ALERT_KEY = 'GitLabPrivateScopeAlert'
  * A global alert telling authenticated users if they need to update GitHub code
  * host token to access the private repositories.
  */
-export const GitHubScopeAlert: FunctionComponent<Props> = ({ authenticatedUser }) => {
+export const CodeHostScopeAlerts: FunctionComponent<Props> = ({ authenticatedUser }) => {
     const { scopes } = useCodeHostScopeContext()
 
     if (!authenticatedUser || scopes === null) {
